BrightSpark-main/ChatBot/requirements.txt
langchain
streamlit
pinecone-client
langchain_core
langchain_experimental
sentence-transformers
openai
langchain_openai
transformers
google-cloud-aiplatform
google-auth
python-dotenv
mysql
apache-airflow-providers-google
mysqlclient
sqlalchemy
psycopg2-binary
pkgconfig
PyPDF2
pypdf
oauth2client
streamlit_float
audio_recorder_streamlit